Get rid of dependency to k8s.io/kubernetes in Telepresence's go.mod
The requirement on `k8s.io/kubernetes` was only used by a unit test in
the `dnet` package, and only by a mock http server in that test. Since
the requirement must be patched, and that patch causes problems in
downstream components, the mock server was moved into its own binary
with its own `go.mod` so that the requirement could be removed from the
telepresence `go.mod` file.

See https://news.ycombinator.com/item?id=27827389 for more info about
why `k8s.io/kubernetes`'s go.mod causes trouble.
